Improve the indexers a bit
