+++ bika/lims/catalog/indexers/__init__.py
@@ -18,11 +18,14 @@
 # Copyright 2018-2019 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-from bika.lims import api
-from plone.indexer import indexer
 from Products.Archetypes.interfaces import IBaseObject
-from Products.CMFPlone.utils import safe_callable
 from Products.CMFPlone.CatalogTool import sortable_title as plone_sortable_title
+from Products.CMFPlone.utils import safe_callable
+from plone.indexer import indexer
+
+from bika.lims import api
+from bika.lims.catalog.bika_catalog import BIKA_CATALOG
+from bika.lims.interfaces import IBikaCatalog
 
 
 @indexer(IBaseObject)
@@ -55,3 +58,24 @@ def sortable_sortkey_title(instance):
         sort_key = 999999
 
     return "{:010.3f}{}".format(sort_key, title)
+
+
+@indexer(IBaseObject, IBikaCatalog)
+def listing_searchable_text(instance):
+    """ Retrieves all the values of metadata columns in the catalog for
+    wildcard searches
+    :return: all metadata values joined in a string
+    """
+    entries = set()
+    catalog = api.get_tool(BIKA_CATALOG)
+    metadata = catalog.getMetadataForUID(api.get_path(instance))
+    for key, brain_value in metadata.items():
+        instance_value = api.safe_getattr(instance, key, None)
+        parsed = api.to_searchable_text_metadata(brain_value or instance_value)
+        entries.add(parsed)
+
+    # Remove empties
+    entries = filter(None, entries)
+
+    # Concatenate all strings to one text blob
+    return " ".join(entries)
