+++ bika/lims/catalog/indexers/configure.zcml
@@ -8,6 +8,9 @@
   <!-- Default case-insenstive sortable_title -->
   <adapter name="sortable_title" factory=".sortable_title"/>
 
+  <!-- Default listing_searchable_text index adapter for bika_catalog -->
+  <adapter name="listing_searchable_text" factory=".listing_searchable_text"/>
+
   <!-- Case-insenstive + sortkey sortable_title for Categories and Analyses -->
   <adapter name="sortable_title" factory=".baseanalysis.sortable_title"/>
   <adapter name="sortable_title" factory=".analysiscategory.sortable_title"/>
@@ -30,7 +33,5 @@
   <adapter name="snapshot_created" factory=".auditlog.snapshot_created"/>
   <adapter name="snapshot_version" factory=".auditlog.snapshot_version"/>
 
-  <!-- Bika Catalog Index Adapters -->
-  <adapter name="listing_searchable_text" factory=".batch.listing_searchable_text"/>
 
 </configure>
